+++ src/lib.rs
@@ -63,7 +63,7 @@ use std::{fmt, io};
 
 use crate::config::{GooseConfiguration, GooseDefaults};
 use crate::controller::{ControllerProtocol, ControllerRequest};
-use crate::goose::{GooseUser, GooseUserCommand, Scenario, Transaction};
+use crate::goose::{GooseUser, GooseUserCommand, Scenario, Transaction, TransactionName};
 use crate::graph::GraphData;
 use crate::logger::{GooseLoggerJoinHandle, GooseLoggerTx};
 use crate::metrics::{GooseMetric, GooseMetrics};
@@ -131,7 +131,7 @@ pub fn is_killswitch_triggered() -> bool {
 }
 
 /// Internal representation of a weighted transaction list.
-type WeightedTransactions = Vec<(usize, String)>;
+type WeightedTransactions = Vec<(usize, TransactionName)>;
 
 /// Internal representation of unsequenced transactions.
 type UnsequencedTransactions = Vec<Transaction>;
@@ -928,7 +928,8 @@ impl GooseAttack {
                 for transaction in scenario.transactions {
                     println!(
                         "    o {} (weight: {})",
-                        transaction.name, transaction.weight
+                        transaction.name.name_for_transaction(),
+                        transaction.weight
                     );
                 }
             }
@@ -2031,19 +2032,19 @@ fn allocate_transactions(
     for transaction in scheduled_sequenced_on_start_transactions.iter() {
         on_start_transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
     for transaction in scheduled_sequenced_transactions.iter() {
         transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
     for transaction in scheduled_sequenced_on_stop_transactions.iter() {
         on_stop_transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
 
@@ -2051,19 +2052,19 @@ fn allocate_transactions(
     for transaction in scheduled_unsequenced_on_start_transactions.iter() {
         on_start_transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
     for transaction in scheduled_unsequenced_transactions.iter() {
         transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
     for transaction in scheduled_unsequenced_on_stop_transactions.iter() {
         on_stop_transactions.extend(vec![(
             *transaction,
-            scenario.transactions[*transaction].name.to_string(),
+            scenario.transactions[*transaction].name.clone(),
         )])
     }
 
@@ -2085,7 +2086,7 @@ fn weight_unsequenced_transactions(
         trace!(
             "{}: {} has weight of {} (reduced with gcd to {})",
             transaction.transactions_index,
-            transaction.name,
+            transaction.name.name_for_transaction(),
             transaction.weight,
             weight
         );